chore(stories): co-locate TokenExtensionBadge[s] with TokenExtensionsStatusRow under app/components/account/token-extensions so all three appear grouped in the Storybook sidebar
