+++ app/components/account/token-extensions/__stories__/TokenExtensionBadges.stories.tsx
@@ -1,21 +1,16 @@
-/* eslint-disable no-restricted-syntax -- test assertions use RegExp for pattern matching */
 import type { Meta, StoryObj } from '@storybook/react';
-import { expect, fn, within } from 'storybook/test';
+import { expect, within } from 'storybook/test';
 
 import * as mockExtensions from '@/app/__tests__/mock-parsed-extensions-stubs';
 import { populatePartialParsedTokenExtension } from '@/app/utils/token-extension';
 
-import { TokenExtensionBadge } from '../TokenExtensionBadge';
+import { TokenExtensionBadges } from '../TokenExtensionBadges';
 
-// More on how to set up stories at: https://storybook.js.org/docs/writing-stories#default-export
 const meta = {
-    args: {
-        onClick: fn(),
-    },
-    component: TokenExtensionBadge,
+    component: TokenExtensionBadges,
     tags: ['autodocs', 'test'],
-    title: 'Components/Common/TokenExtensionBadge',
-} satisfies Meta<typeof TokenExtensionBadge>;
+    title: 'Components/Account/token-extensions/TokenExtensionBadges',
+} satisfies Meta<typeof TokenExtensionBadges>;
 
 export default meta;
 type Story = StoryObj<typeof meta>;
@@ -29,12 +24,11 @@ const extension = {
 // More on writing stories with args: https://storybook.js.org/docs/writing-stories/args
 export const Primary: Story = {
     args: {
-        extension,
+        extensions: new Array(5).fill(null).map(() => extension),
     },
     async play({ canvasElement }) {
         const canvas = within(canvasElement);
-        // The badge shows "Enabled" or "Disabled" by default, not the extension name
-        const badge = await canvas.findByText(/Enabled|Disabled/);
-        expect(badge).toBeInTheDocument();
+        const badges = canvas.getAllByText('transferFeeConfig');
+        expect(badges).toHaveLength(5);
     },
 };
